Move DealWithFees to cli::impls
